Renamed components to packages Changed the accessibility and naming of some of the methods in RecursiveParameterRequestCommand. Made Jenkinfile selection on package creation possible.
