+++ src/atomist.config.ts
@@ -16,7 +16,7 @@ import {TeamMemberCreated} from "./gluon/member/TeamMemberCreated";
 import {TeamMemberCreatedEvent} from "./gluon/member/teamMemberIngester";
 import {ApplicationCreated} from "./gluon/packages/ApplicationCreated";
 import {ApplicationCreatedEvent} from "./gluon/packages/applicationsIngester";
-import {ConfigureComponent} from "./gluon/packages/ConfigureComponent";
+import {ConfigurePackage} from "./gluon/packages/ConfigurePackage";
 import {
     CreateApplication,
     LinkExistingApplication,
@@ -103,7 +103,7 @@ export const configuration: any = {
         AddConfigServer,
         ListTeamProjects,
         ListProjectDetails,
-        ConfigureComponent,
+        ConfigurePackage,
     ],
     events: [
         TeamCreated,
