+++ src/gluon/packages/ApplicationCreated.ts
@@ -10,7 +10,7 @@ import {
 import {buttonForCommand} from "@atomist/automation-client/spi/message/MessageClient";
 import {url} from "@atomist/slack-messages";
 import {QMConfig} from "../../config/QMConfig";
-import {ConfigureComponent} from "./ConfigureComponent";
+import {ConfigurePackage} from "./ConfigurePackage";
 
 @EventHandler("Receive ApplicationCreatedEvent events", `
 subscription ApplicationCreatedEvent {
@@ -83,7 +83,7 @@ export class ApplicationCreated implements HandleEvent<any> {
                     actions: [
                         buttonForCommand(
                             {text: "Configure Component"},
-                            new ConfigureComponent(),
+                            new ConfigurePackage(),
                             {
                                 projectName : applicationCreatedEvent.project.name,
                                 applicationName : applicationCreatedEvent.application.name,
